+++ src/Concerns/AddsIncludesToQuery.php
@@ -34,26 +34,19 @@ trait AddsIncludesToQuery
     protected function addRequestedIncludesToQuery()
     {
         $this->request->includes()
-            ->flatMap(function (string $include) {
-                $relatedTables = collect(explode('.', $include));
-
-                return $relatedTables
-                    ->mapWithKeys(function ($table, $key) use ($relatedTables) {
-                        $fields = $this->getRequestedFieldsForRelatedTable(Str::snake($table));
-
-                        $fullRelationName = $relatedTables->slice(0, $key + 1)->implode('.');
-
-                        if (empty($fields)) {
-                            return [$fullRelationName];
-                        }
-
-                        return [$fullRelationName => function ($query) use ($fields) {
-                            $query->select($fields);
-                        }];
-                    });
+            ->reject(function (string $include) {
+                return Str::endsWith($include, config('query-builder.count_suffix'));
             })
-            ->pipe(function (Collection $withs) {
-                $this->with($withs->all());
+            ->pipe(function (Collection $includes) {
+                $this->addIncludedWithsToQuery($includes);
+            });
+
+        $this->request->includes()
+            ->filter(function (string $include) {
+                return Str::endsWith($include, config('query-builder.count_suffix'));
+            })
+            ->pipe(function (Collection $withCounts) {
+                $this->addIncludedWithCountsToQuery($withCounts);
             });
     }
 
@@ -80,5 +73,40 @@ trait AddsIncludesToQuery
 
                 return $includes->push("{$includes->last()}.{$relationship}");
             }, collect());
+    }
+
+    protected function addIncludedWithsToQuery(Collection $includes)
+    {
+        $includes
+            ->flatMap(function (string $include) {
+                $relatedTables = collect(explode('.', $include));
+
+                return $relatedTables
+                    ->mapWithKeys(function ($table, $key) use ($relatedTables) {
+                        $fields = $this->getRequestedFieldsForRelatedTable(Str::snake($table));
+
+                        $fullRelationName = $relatedTables->slice(0, $key + 1)->implode('.');
+
+                        if (empty($fields)) {
+                            return [$fullRelationName];
+                        }
+
+                        return [$fullRelationName => function ($query) use ($fields) {
+                            $query->select($fields);
+                        }];
+                    });
+            })
+            ->pipe(function (Collection $withs) {
+                $this->with($withs->all());
+            });
+    }
+
+    protected function addIncludedWithCountsToQuery(Collection $includes)
+    {
+        $counts = $includes->map(function (string $include) {
+            return Str::before($include, config('query-builder.count_suffix'));
+        });
+
+        $this->withCount($counts->toArray());
     }
 }