+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to `laravel-query-builder` will be documented in this file
 - renamed `Spatie\QueryBuilder\Sort` to `Spatie\QueryBuilder\AllowedSort`
 - renamed `Spatie\QueryBuilder\Included` to `Spatie\QueryBuilder\AllowedInclude`
 - renamed `Spatie\QueryBuilder\Filter` to `Spatie\QueryBuilder\AllowedFilter`
+- `Filter`, `Include` and `Sort` get the entire `QueryBuilder` instance to manipulate
+- `Filter`, `Include` and `Sort` interfaces no longer need to return the `Builder` instance
 
 ## 1.17.3 - 2019-04-16